+++ Parrot/Patches/ZNetPatch.cs
@@ -5,6 +5,8 @@ using System.Reflection.Emit;
 
 using HarmonyLib;
 
+using static PluginConfig;
+
 [HarmonyPatch(typeof(ZNet))]
 static class ZNetPatch {
   [HarmonyTranspiler]
@@ -17,18 +19,18 @@ static class ZNetPatch {
             new CodeMatch(OpCodes.Ldarg_0),
             new CodeMatch(OpCodes.Ldloc_1),
             new CodeMatch(OpCodes.Call, AccessTools.Method(typeof(ZNet), nameof(ZNet.OnNewConnection))))
-        .ThrowIfInvalid("Could not patch ZNet.CheckForIncommingServerConnections()! (OnNewConnection)")
+        .ThrowIfInvalid("Could not patch ZNet.CheckForIncommingServerConnections()! (on-new-connection)")
         .CreateLabel(out Label onNewConnectionLabel)
         .InsertAndAdvance(
             new CodeInstruction(OpCodes.Ldloc_1),
-            Transpilers.EmitDelegate(OnNewConnectionDelegate),
+            new CodeInstruction(OpCodes.Call, AccessTools.Method(typeof(ZNetPatch), nameof(OnNewConnectionDelegate))),
             new CodeInstruction(OpCodes.Brfalse, onNewConnectionLabel),
             new CodeInstruction(OpCodes.Ret))
         .InstructionEnumeration();
   }
 
   static bool OnNewConnectionDelegate(ZNetPeer netPeer) {
-    if (ConnectionManager.IsSteamGameServer(netPeer)) {
+    if (AllowParrotServerConnections.Value && ConnectionManager.IsSteamGameServer(netPeer)) {
       ConnectionManager.RegisterParrotClient(netPeer);
       return true;
     }
@@ -40,5 +42,32 @@ static class ZNetPatch {
   [HarmonyPatch(nameof(ZNet.Disconnect))]
   static void DisconnectPrefix(ZNetPeer peer) {
     ConnectionManager.ParrotNetPeers.Remove(peer);
+  }
+
+  [HarmonyTranspiler]
+  [HarmonyPatch(nameof(ZNet.SendPlayerList))]
+  static IEnumerable<CodeInstruction> SendPlayerListTranspiler(
+      IEnumerable<CodeInstruction> instructions, ILGenerator generator) {
+    return new CodeMatcher(instructions, generator)
+        .Start()
+        .MatchStartForward(
+            new CodeMatch(OpCodes.Ldfld, AccessTools.Field(typeof(ZNet), nameof(ZNet.m_players))),
+            new CodeMatch(OpCodes.Callvirt),
+            new CodeMatch(
+                OpCodes.Callvirt, AccessTools.Method(typeof(ZPackage), nameof(ZPackage.Write), [typeof(int)])))
+        .ThrowIfInvalid($"Could not patch ZNet.SendPlayerList()! (write-players-count)")
+        .Advance(offset: 3)
+        .InsertAndAdvance(
+            new CodeInstruction(OpCodes.Ldarg_0),
+            new CodeInstruction(OpCodes.Ldloc_0),
+            new CodeInstruction(
+                OpCodes.Call, AccessTools.Method(typeof(ZNetManager), nameof(ZNetManager.AddServerPlayers))))
+        .InstructionEnumeration();
+  }
+
+  [HarmonyPostfix]
+  [HarmonyPatch(nameof(ZNet.Start))]
+  static void StartPostfix(ZNet __instance) {
+    ZNetManager.SetupServerPlayer(__instance);
   }
 }